# Task3/arkpz-pzpi-22-3-shut-oleksandr-task3/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/rooms": {
      "get": {
        "operationId": "RoomsController_getAllRooms",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Room"
                  }
                }
              }
            }
          }
        },
        "summary": "Get all rooms",
        "tags": [
          "rooms"
        ]
      },
      "post": {
        "operationId": "RoomsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRoomDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Create room",
        "tags": [
          "rooms"
        ]
      },
      "patch": {
        "operationId": "RoomsController_update",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRoomDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Update room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/rooms/{id}": {
      "get": {
        "operationId": "RoomsController_getRoomById",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Get room by id",
        "tags": [
          "rooms"
        ]
      },
      "delete": {
        "operationId": "RoomsController_delete",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "summary": "Delete room",
        "tags": [
          "rooms"
        ]
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginResponseDto"
                }
              }
            }
          }
        },
        "summary": "user login",
        "tags": [
          "auth"
        ]
      }
    },
    "/api/auth/registration": {
      "post": {
        "operationId": "AuthController_registration",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterResponseDto"
                }
              }
            }
          }
        },
        "summary": "user register",
        "tags": [
          "auth"
        ]
      }
    },
...
    "/api/sensors-adapter/temperature/{id}": {
      "post": {
        "operationId": "SensorsAdapterController_postTemperature",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Post temperature data by room id",
        "tags": [
          "sensors-adapter"
        ]
      }
    },
    "/api/sensors-adapter/moisture/{id}": {
      "post": {
        "operationId": "SensorsAdapterController_postMoisture",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Post moisture data by room id",
        "tags": [
          "sensors-adapter"
        ]
      }
    },
    "/api/sensors-adapter/carbonDioxide/{id}": {
      "post": {
        "operationId": "SensorsAdapterController_postCarbonDioxide",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Post carbon dioxide data by room id",
        "tags": [
          "sensors-adapter"
        ]
      }
    },
    "/api/sensors-adapter/illumination/{id}": {
      "post": {
        "operationId": "SensorsAdapterController_postIllumination",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Post illumination data by room id",
        "tags": [
          "sensors-adapter"
        ]
      }
    },
    "/api/illumination-adapter/{id}": {
      "get": {
        "operationId": "IlluminationAdapterController_getIllumination",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "number"
                }
              }
            }
          }
        },
        "summary": "Get illumination data by room id",
        "tags": [
          "illumination-adapter"
        ]
      },
      "post": {
        "operationId": "IlluminationAdapterController_postIllumination",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSensorDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Room"
                }
              }
            }
          }
        },
        "summary": "Post illumination data by room id",
        "tags": [
          "illumination-adapter"
        ]
      }
    }
  },
  "info": {
    "title": "Project api",
    "description": "Here stores api documentation",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "username": {
            "type": "string",
            "example": "cdidk123",
            "description": "username"
          },
          "password": {
            "type": "string",
            "example": "VERY_STRONG_PASSWORD",
            "description": "user password"
          }
        },
        "required": [
          "username",
          "password"
        ]
      },
...
      "Room": {
        "type": "object",
        "properties": {}
      },
      "CreateRoomDto": {
        "type": "object",
        "properties": {
          "user": {
            "type": "string",
            "example": "userId",
            "description": "User ID"
          },
          "temperature": {
            "type": "number",
            "example": 22,
            "description": "Temperature"
          },
          "moisture": {
            "type": "number",
            "example": 50,
            "description": "Moisture"
          },
          "carbonDioxide": {
            "type": "number",
            "example": 400,
            "description": "Carbon Dioxide (CO2)"
          },
          "illumination": {
            "type": "number",
            "example": 300,
            "description": "Illumination"
          }
        },
        "required": [
          "user",
          "temperature",
          "moisture",
          "carbonDioxide",
          "illumination"
        ]
      },
      "UpdateRoomDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "roomID",
            "description": "Room ID"
          },
          "user": {
            "type": "string",
            "example": "userId",
            "description": "User ID"
          },
          "temperature": {
            "type": "number",
            "example": 22,
            "description": "Temperature"
          },
          "moisture": {
            "type": "number",
            "example": 50,
            "description": "Moisture"
          },
          "carbonDioxide": {
            "type": "number",
            "example": 400,
            "description": "Carbon Dioxide (CO2)"
          },
          "illumination": {
            "type": "number",
            "example": 300,
            "description": "Illumination"
          }
        },
        "required": [
          "id",
          "user",
          "temperature",
          "moisture",
          "carbonDioxide",
          "illumination"
        ]
      },
      "LoginResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Login successful",
            "description": "Response message"
          },
          "token": {
            "type": "object",
            "example": {
              "accessToken": "yourAccessToken"
            },
            "description": "Token object"
          }
        },
        "required": [
          "message",
          "token"
        ]
      },
      "RegisterResponseDto": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "example": "Register successful",
            "description": "Response message"
          },
          "token": {
            "type": "object",
            "example": {
              "accessToken": "yourAccessToken"
            },
            "description": "Token object"
          }
        },
        "required": [
          "message",
          "token"
        ]
      },
...
      "UpdateSensorDto": {
        "type": "object",
        "properties": {
          "value": {
            "type": "number",
            "example": 300,
            "description": "Sensor value"
          }
        },
        "required": [
          "value"
        ]
      }
    }
  }
}